# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ArogyaGenie Healthcare Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
    put:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/me/onboard:
    post:
      operationId: onboardUser
      tags: [users]
      summary: Complete user onboarding (set role and basic info)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingInput"
      responses:
        "200":
          description: Onboarded user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  # ─── DOCTORS ────────────────────────────────────────────────────────────────
...
  /prescriptions:
    get:
      operationId: listPrescriptions
      tags: [prescriptions]
      summary: List prescriptions for current user
      responses:
        "200":
          description: List of prescriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prescription"
    post:
      operationId: createPrescription
      tags: [prescriptions]
      summary: Create a prescription (doctor only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrescriptionInput"
      responses:
        "201":
          description: Created prescription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prescription"
  /prescriptions/{id}:
    get:
      operationId: getPrescription
      tags: [prescriptions]
      summary: Get prescription by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Prescription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prescription"
    patch:
      operationId: updatePrescription
      tags: [prescriptions]
      summary: Update prescription status or details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrescriptionUpdate"
      responses:
        "200":
          description: Updated prescription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prescription"

  # ─── LAB REPORTS ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: number
        clerkId:
          type: string
        email:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        userName:
          type: ["string", "null"]
        name:
          type: ["string", "null"]
        displayName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [patient, doctor, diagnostic_center, pharmacy, admin]
        status:
          type: string
          enum: [pending, active, suspended]
        avatarUrl:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        age:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        bloodGroup:
          type: ["string", "null"]
        allergies:
          type: ["string", "null"]
        existingConditions:
          type: ["string", "null"]
        currentMedications:
          type: ["string", "null"]
        previousIllnesses:
          type: ["string", "null"]
        emergencyContact:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - clerkId
        - email
        - role
        - status
        - createdAt
    UserProfileUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
        age:
          type: string
        gender:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        bloodGroup:
          type: string
        allergies:
          type: string
        existingConditions:
          type: string
        currentMedications:
          type: string
        previousIllnesses:
          type: string
        emergencyContact:
          type: string
        avatarUrl:
          type: string
    OnboardingInput:
      type: object
      required:
        - role
        - firstName
        - lastName
      properties:
        role:
          type: string
          enum: [patient, doctor, diagnostic_center, pharmacy]
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
...
    Prescription:
      type: object
      properties:
        id:
          type: number
        patientId:
          type: number
        doctorId:
          type: number
        appointmentId:
          type: ["number", "null"]
        patientName:
          type: ["string", "null"]
        doctorName:
          type: ["string", "null"]
        medicines:
          type: string
        diagnosis:
          type: ["string", "null"]
        instructions:
          type: ["string", "null"]
        fileUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, dispensed, expired]
        prescribedDate:
          type: string
        createdAt:
          type: string
      required:
        - id
        - patientId
        - doctorId
        - medicines
        - status
        - prescribedDate
        - createdAt
    PrescriptionInput:
      type: object
      required:
        - patientId
        - medicines
        - prescribedDate
      properties:
        patientId:
          type: number
        appointmentId:
          type: number
        medicines:
          type: string
        diagnosis:
          type: string
        instructions:
          type: string
        fileUrl:
          type: string
        prescribedDate:
          type: string
    PrescriptionUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [active, dispensed, expired]
        diagnosis:
          type: string
        medicines:
          type: string
        instructions:
          type: string

    # ─── LAB REPORT ────────────────────────────────────────────────────────────
...